Add nrows and size to table metadata and tap

+++ daiquiri/query/models.py
@@ -69,8 +69,8 @@ class QueryJob(Job):
     actual_query = models.TextField(null=True, blank=True)
 
     queue = models.CharField(max_length=16, null=True, blank=True)
-    nrows = models.IntegerField(null=True, blank=True)
-    size = models.IntegerField(null=True, blank=True)
+    nrows = models.BigIntegerField(null=True, blank=True)
+    size = models.BigIntegerField(null=True, blank=True)
 
     metadata = JSONField()
